+++ tests/testthat/test-basic.r
@@ -113,4 +113,5 @@ test_that('partial name causes error', {
 test_that('trailing comma is accepted', {
     expect_error(box::use(mod/a, ), NA)
     expect_error(box::use(mod/a, mod/b, ), NA)
+    expect_error(box::use(mod/a[modname, double, ]), NA)
 })